Completes ridding of OWNED_MODE
